# cmd/api/docs/swagger.yaml
definitions:
...
  models.UpdateInventory:
    properties:
      category_id:
        type: integer
      description:
        type: string
      price:
        type: number
      productName:
        type: string
      stock:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/inventories/add:
    post:
      consumes:
      - multipart/form-data
      description: Admin can add new  products
      parameters:
      - description: category_id
        in: formData
        name: category_id
        required: true
        type: string
      - description: product_name
        in: formData
        name: product_name
        required: true
        type: string
      - description: description
        in: formData
        name: description
        required: true
        type: string
      - description: price
        in: formData
        name: price
        required: true
        type: string
      - description: stock
        in: formData
        name: stock
        required: true
        type: string
      - description: image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Inventory
      tags:
      - Admin
  /admin/inventories/add-image:
    post:
      consumes:
      - multipart/form-data
      description: Admin can add new image to product
      parameters:
      - description: product_id
        in: formData
        name: product_id
        required: true
        type: string
      - description: image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add image to an Inventory
      tags:
      - Admin
  /admin/inventories/delete:
    delete:
      consumes:
      - application/json
      description: Admin can delete a product
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Inventory
      tags:
      - Admin
  /admin/inventories/delete-image:
    delete:
      consumes:
      - application/json
      description: Admin can delete a product image
      parameters:
      - description: product_id
        in: query
        name: product_id
        required: true
        type: string
      - description: image_id
        in: query
        name: image_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Inventory image
      tags:
      - Admin
  /admin/inventories/update:
    patch:
      consumes:
      - application/json
      description: Admin can update stock of the inventories
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: Update Inventory
        in: body
        name: updateinventory
        required: true
        schema:
          $ref: '#/definitions/models.UpdateInventory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Stock
      tags:
      - Admin
...
  /admin/products:
    get:
      consumes:
      - application/json
      description: client can view the list of available products
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Productss
      tags:
      - Admin
...
  /products:
    get:
      consumes:
      - application/json
      description: client can view the list of available products
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Products
      tags:
      - Products
  /products/category:
    get:
      consumes:
      - application/json
      description: client can filter with a category and get the list of  products
        in the category
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: filter Products by category
      tags:
      - Products
  /products/category_details:
    get:
      consumes:
      - application/json
      description: client can view the list of available products
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: id
        in: query
        name: catid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Products based on category
      tags:
      - Products
  /products/details:
    get:
      consumes:
      - application/json
      description: client can view the details of the product
      parameters:
      - description: Inventory ID
        in: query
        name: inventoryID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Show Product Details
      tags:
      - Products
  /products/search:
    get:
      consumes:
      - application/json
      description: client can search with a key and get the list of products similar
        to that key
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      - description: searchkey
        in: query
        name: searchkey
        required: true
        type: string
      - description: sortBY (asc/desc) - Sort by price in ascending (asc) or descending
          (desc) order
        in: query
        name: sortBY
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Search Products
      tags:
      - Products
...
swagger: "2.0"